requirements.txt
pyinstaller
pip
wheel
fastapi==0.109.2
prometheus-client==0.17.1
prometheus-fastapi-instrumentator==6.0.0
pydantic==2.1.1
pydantic_core==2.4.0
pytz==2022.1
PyYAML==6.0.1
requests==2.31.0
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.23.2
yarl==1.9.2

pytest==7.4.4

tabulate
psutil==5.9.8

python-multipart

sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
greenlet==3.0.3